requirements.txt
flask==3.1.1
ultralytics==8.3.159
opencv-python-headless==4.11.0.86
# torch==2.7.1+cpu
torch==2.7.1
numpy==2.0.2
gunicorn==22.0.0
paho-mqtt==1.6.1
